A work around for the null connection that can sometimes come back from bridge. Like loss of bridge itself, reload after a short delay. It appears the cause the page changing while a reload is happening, the javascript side is not terminated. Handle situation where switching from system dbus broker to embedded (only really happens in dev environment).

+++ client-logonbox-vpn-gui-jfx/src/main/java/com/logonbox/vpn/client/gui/jfx/UI.java
@@ -1753,11 +1753,13 @@ public class UI implements BusLifecycleListener {
 				}
 				logoFile = null;
 			}
-			try {
-				Files.delete(splashFile);
-			}
-			catch(IOException ioe) {
-				log.warn("Failed to delete.", ioe);
+			if(Files.exists(splashFile)) {
+				try {
+					Files.delete(splashFile);
+				}
+				catch(IOException ioe) {
+					log.warn("Failed to delete.", ioe);
+				}
 			}
 			updateVMOptions(null);
 		} else {
